refactor(notebooks): Move viewset to presentation, wire tach interface
Generated-By: PostHog Code
Task-Id: f29c0939-1cf9-4268-9ef9-b252d992ea55
