+++ products/notebooks/backend/presentation/__init__.py
@@ -1,0 +1,1 @@
+"""HTTP presentation layer of the notebooks product (DRF viewsets and serializers)."""
